requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-environ
requests
google-auth